Update README.md
This project aims to create a virtual application development environment with Docker containers on AWS Cloud. This involves provisioning an EC2 instance, installing Docker, configuring a Docker Compose file, building and deploying the containers, and testing the application. This approach offers benefits such as flexibility, portability, scalability, and simplified deployment.